Fix unportable coding in BackgroundWorkerStateChange().
PIDs aren't necessarily ints; our usual practice for printing them
is to explicitly cast to long.  Per buildfarm member rover_firefly.

+++ src/backend/postmaster/bgworker.c
@@ -303,8 +303,8 @@ BackgroundWorkerStateChange(void)
 		rw->rw_worker.bgw_notify_pid = slot->worker.bgw_notify_pid;
 		if (!PostmasterMarkPIDForWorkerNotify(rw->rw_worker.bgw_notify_pid))
 		{
-			elog(DEBUG1, "worker notification PID %u is not valid",
-				 rw->rw_worker.bgw_notify_pid);
+			elog(DEBUG1, "worker notification PID %lu is not valid",
+				 (long) rw->rw_worker.bgw_notify_pid);
 			rw->rw_worker.bgw_notify_pid = 0;
 		}
